Reorganize components.
