+++ example/src/components/atoms/Checkbox.styl
