+++ example/src/components/molecules/Example.styl
