+++ example/src/components/molecules/InputGroup.js
@@ -1,7 +1,7 @@
 import React, { PureComponent } from 'react'
 import PropTypes from 'prop-types'
 
-import Button from './Button'
+import { Button } from '@atoms'
 import './InputGroup.styl'
 
 class InputGroup extends PureComponent {
