+++ example/src/components/molecules/index.js
@@ -1,4 +1,3 @@
-export { default as Button } from './Button'
-export { default as Checkbox } from './Checkbox'
+export { default as Example } from './Example'
 export { default as FileButton } from './FileButton'
 export { default as InputGroup } from './InputGroup'
