+++ example/src/components/organisms/Footer.js
@@ -1,6 +1,6 @@
 import React, { PureComponent } from 'react'
 
-import { Button } from './common'
+import { Button } from '@atoms'
 
 class Footer extends PureComponent {
   static propTypes = {};
