+++ example/src/components/organisms/Storage.styl
