+++ example/src/components/organisms/index.js
@@ -1,3 +1,5 @@
 export { default as AuthenticationExample } from './Authentication'
+export { default as Footer } from './Footer'
+export { default as Header } from './Header'
 export { default as StorageExample } from './Storage'
 export { default as TodoListExample } from './TodoList'
