+++ example/webpack.config.js
@@ -5,7 +5,10 @@ module.exports = {
   entry: ['babel-polyfill', './index.js'],
   resolve: {
     alias: {
-      '@assets': path.resolve(__dirname, 'src/assets')
+      '@assets': path.resolve(__dirname, 'src/assets'),
+      '@atoms': path.resolve(__dirname, 'src/components/atoms'),
+      '@molecules': path.resolve(__dirname, 'src/components/molecules'),
+      '@organisms': path.resolve(__dirname, 'src/components/organisms')
     }
   },
   module: {
